vagrant: add Vagrantfile for clang + ASan&UBSan
Run only the unit tests so far, until the issue with pre-loading
of the clang-specific ASan DSO is resolved

See: https://github.com/systemd/systemd-centos-ci/pull/114#issuecomment-495399367

+++ vagrant/vagrant-test-sanitizers.sh
@@ -31,19 +31,24 @@ export QEMU_SMP=$(nproc)
 # are compiled in as modules
 export SKIP_INITRD=no
 
-# 1) Run it under systemd-nspawn
-rm -fr /var/tmp/systemd-test*
-exectask "TEST-01-BASIC_sanitizers-nspawn" "make -C test/TEST-01-BASIC clean setup run clean-again TEST_NO_QEMU=1"
-NSPAWN_EC=$?
-# Each integration test dumps the system journal when something breaks
-rsync -amq /var/tmp/systemd-test*/journal "$LOGDIR/TEST-01-BASIC_sanitizers-nspawn/" || :
-
-if [[ $NSPAWN_EC -eq 0 ]]; then
-    # 2) Run it under QEMU, but only if the systemd-nspawn run was successful
+# Temporarily disable this part until the clang-specific DSO preloading issues
+# are dealth with, see https://github.com/systemd/systemd-centos-ci/pull/114#issuecomment-495399367
+# However, still run this with gcc & libasan
+if ldd build/systemd | grep -q libasan; then
+    # 1) Run it under systemd-nspawn
     rm -fr /var/tmp/systemd-test*
-    exectask "TEST-01-BASIC_sanitizers-qemu" "make -C test/TEST-01-BASIC clean setup run clean-again TEST_NO_NSPAWN=1"
+    exectask "TEST-01-BASIC_sanitizers-nspawn" "make -C test/TEST-01-BASIC clean setup run clean-again TEST_NO_QEMU=1"
+    NSPAWN_EC=$?
     # Each integration test dumps the system journal when something breaks
-    rsync -amq /var/tmp/systemd-test*/journal "$LOGDIR/TEST-01-BASIC_sanitizers-qemu/" || :
+    rsync -amq /var/tmp/systemd-test*/journal "$LOGDIR/TEST-01-BASIC_sanitizers-nspawn/" || :
+
+    if [[ $NSPAWN_EC -eq 0 ]]; then
+        # 2) Run it under QEMU, but only if the systemd-nspawn run was successful
+        rm -fr /var/tmp/systemd-test*
+        exectask "TEST-01-BASIC_sanitizers-qemu" "make -C test/TEST-01-BASIC clean setup run clean-again TEST_NO_NSPAWN=1"
+        # Each integration test dumps the system journal when something breaks
+        rsync -amq /var/tmp/systemd-test*/journal "$LOGDIR/TEST-01-BASIC_sanitizers-qemu/" || :
+    fi
 fi
 
 # Summary
